ProxyServer выделен отдельным файлом. Изменены дефолтные настройки запуска. Blacklist перечитывается непосредственно перед проверкой, поэтому вносить изменения в список можно без перезагрузки программы.
